Refactor of file structure
