--- src/main/java/ru/spbu/apcyb/svp/tasks/task2/Task2.java
@@ -1,11 +1,0 @@
-package ru.spbu.apcyb.svp.tasks.task2;
-
-/**
- * Задание 2.
- */
-public class Task2 {
-
-  public static void main(String[] args) {
-    int x = 0; // placeholder
-  }
-}
