+++ src/test/java/ru/spbu/apcyb/svp/tasks/task2/Task2Test.java
@@ -1,4 +1,4 @@
-package ru.spbu.apcyb.svp.tasks;
+package ru.spbu.apcyb.svp.tasks.task2;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
